refactor(renderer): extract live-agent index patch into worktree-agent-live-index-patch module (max-lines)
